+++ src/providers/twitch/pubsubmessages/ChannelPoints.hpp
@@ -19,17 +19,7 @@ struct PubSubCommunityPointsChannelV1Message {
 
     QJsonObject data;
 
-    PubSubCommunityPointsChannelV1Message(const QJsonObject &root)
-        : typeString(root.value("type").toString())
-        , data(root.value("data").toObject())
-    {
-        auto oType =
-            magic_enum::enum_cast<Type>(this->typeString.toStdString());
-        if (oType.has_value())
-        {
-            this->type = oType.value();
-        }
-    }
+    PubSubCommunityPointsChannelV1Message(const QJsonObject &root);
 };
 
 }  // namespace chatterino
