+++ src/providers/twitch/pubsubmessages/ChatModeratorAction.hpp
@@ -20,17 +20,7 @@ struct PubSubChatModeratorActionMessage {
 
     QJsonObject data;
 
-    PubSubChatModeratorActionMessage(const QJsonObject &root)
-        : typeString(root.value("type").toString())
-        , data(root.value("data").toObject())
-    {
-        auto oType =
-            magic_enum::enum_cast<Type>(this->typeString.toStdString());
-        if (oType.has_value())
-        {
-            this->type = oType.value();
-        }
-    }
+    PubSubChatModeratorActionMessage(const QJsonObject &root);
 };
 
 }  // namespace chatterino
